+++ src/index.js
@@ -221,7 +221,7 @@ class Service extends AdapterService {
         return query[method].call(query, column, value); // eslint-disable-line no-useless-call
       }
 
-      const property = this.jsonSchema && (this.jsonSchema.properties[column] || (methodKey && this.jsonSchema.properties[methodKey]));
+      const property = this.jsonSchema && this.jsonSchema.properties && (this.jsonSchema.properties[column] || (methodKey && this.jsonSchema.properties[methodKey]));
       let columnType = property && property.type;
       if (columnType) {
         if (Array.isArray(columnType)) { columnType = columnType[0]; }
@@ -358,7 +358,7 @@ class Service extends AdapterService {
 
   _selectQuery (q, $select) {
     if ($select && Array.isArray($select)) {
-      const items = $select.concat(`${this.Model.tableName}.${this.id}`); // BUG if this.id is array
+      const items = $select.concat(Array.isArray(this.id) ? this.id.map(el => { return `${this.Model.tableName}.${el}`; }) : `${this.Model.tableName}.${this.id}`);
 
       for (const [key, item] of Object.entries(items)) {
         const matches = item.match(/^ref\((.+)\)( as (.+))?$/);
@@ -519,7 +519,7 @@ class Service extends AdapterService {
    * `find` service function for Objection.
    * @param params
    */
-  _find (params) {
+  _find (params = {}) {
     const find = (params, count, filters, query) => {
       const q = params.objection || this.createQuery(params);
       const groupByColumns = this.getGroupByColumns(q);
@@ -558,7 +558,7 @@ class Service extends AdapterService {
 
       if (count) {
         const countColumns = groupByColumns || (Array.isArray(this.id) ? this.id.map(idKey => `${this.Model.tableName}.${idKey}`) : [`${this.Model.tableName}.${this.id}`]);
-        const countQuery = this._createQuery(params);
+        const countQuery = this._createQuery(params).clear('groupBy');
 
         if (query.$joinRelation) {
           countQuery
@@ -574,7 +574,7 @@ class Service extends AdapterService {
           countQuery.count({ total: countColumns });
         }
 
-        if (query && query.$modify) {
+        if (query && query.$modify && params.modifierFiltersResults !== false) {
           this.modifyQuery(countQuery, query.$modify);
         }
 
@@ -599,7 +599,7 @@ class Service extends AdapterService {
     return result;
   }
 
-  _get (id, params) {
+  _get (id, params = {}) {
     // merge user query with the 'id' to get
     const findQuery = Object.assign({}, { $and: [] }, params.query);
     findQuery.$and.push(this.getIdsQuery(id)); // BUG will fail with composite primary key because table name will be missing
@@ -616,58 +616,107 @@ class Service extends AdapterService {
       });
   }
 
+  _getCreatedRecords (insertResults, inputData, params) {
+    if (params.query && params.query.$noSelect) {
+      return inputData;
+    }
+    if (!Array.isArray(insertResults)) {
+      insertResults = [insertResults];
+    }
+
+    const findQuery = Object.assign({ $and: [] }, params.query);
+    const idsQueries = [];
+
+    if (Array.isArray(this.id)) {
+      for (const insertResult of insertResults) {
+        const ids = [];
+        for (const idKey of this.id) {
+          if (idKey in insertResult) {
+            ids.push(insertResult[idKey]);
+          } else {
+            return inputData;
+          }
+        }
+        idsQueries.push(this.getIdsQuery(ids));
+      }
+    } else {
+      const ids = [];
+      for (const insertResult of insertResults) {
+        if (this.id in insertResult) {
+          ids.push(insertResult[this.id]);
+        } else {
+          return inputData;
+        }
+      }
+      idsQueries.push(this.getIdsQuery(null, ids));
+    }
+
+    if (idsQueries.length > 1) {
+      findQuery.$and.push({ $or: idsQueries });
+    } else {
+      findQuery.$and = findQuery.$and.concat(idsQueries);
+    }
+
+    return this._find(Object.assign({}, params, { query: findQuery }))
+      .then(page => {
+        const records = page.data || page;
+        if (Array.isArray(inputData)) {
+          return records;
+        }
+        return records[0];
+      });
+  }
+
+  /**
+   * @param data
+   * @param params
+   * @returns {Promise<Object|Object[]>}
+   * @private
+   */
+  _batchInsert (data, params) {
+    const { dialect } = this.Model.knex().client;
+    // batch insert only works with Postgresql and SQL Server
+    if (dialect === 'postgresql' || dialect === 'mssql') {
+      return this._createQuery(params)
+        .insert(data)
+        .returning(this.id);
+    }
+    if (!Array.isArray(data)) {
+      return this._createQuery(params).insert(data);
+    }
+    const promises = data.map(dataItem => {
+      return this._createQuery(params).insert(dataItem);
+    });
+    return Promise.all(promises);
+  }
+
   /**
    * `create` service function for Objection.
    * @param {object} data
    * @param {object} params
    */
-  async _create (data, params) {
+  async _create (data, params = {}) {
     const transaction = await this._createTransaction(params);
-    const create = (data, params) => {
-      const q = this._createQuery(params);
-      const allowedUpsert = this.mergeRelations(this.allowedUpsert, params.mergeAllowUpsert);
-      const allowedInsert = this.mergeRelations(this.allowedInsert, params.mergeAllowInsert);
+    const q = this._createQuery(params);
+    let promise = q;
+    const allowedUpsert = this.mergeRelations(this.allowedUpsert, params.mergeAllowUpsert);
+    const allowedInsert = this.mergeRelations(this.allowedInsert, params.mergeAllowInsert);
 
-      if (this.createUseUpsertGraph) {
-        if (allowedUpsert) {
-          q.allowGraph(allowedUpsert);
-        }
-        q.upsertGraphAndFetch(data, this.upsertGraphOptions);
-      } else if (allowedInsert) {
-        q.allowGraph(allowedInsert);
-        q.insertGraph(data, this.insertGraphOptions);
-      } else {
-        q.insert(data, this.id);
+    if (this.createUseUpsertGraph) {
+      if (allowedUpsert) {
+        q.allowGraph(allowedUpsert);
       }
-
-      return q
-        .then(row => {
-          if (params.query && params.query.$noSelect) { return data; }
-
-          let id;
-
-          if (Array.isArray(this.id)) {
-            id = [];
-
-            for (const idKey of this.id) {
-              id.push(row && row[idKey] ? row[idKey] : data[idKey]);
-            }
-          } else {
-            id = row && row[this.id] ? row[this.id] : data[this.id];
-          }
-
-          if (!id || (Array.isArray(id) && !id.length)) { return data; }
-
-          return this._get(id, params);
-        })
-        .catch(errorHandler);
-    };
-
-    if (Array.isArray(data)) {
-      return Promise.all(data.map(current => create(current, params))).then(this._commitTransaction(transaction), this._rollbackTransaction(transaction));
+      q.upsertGraph(data, this.upsertGraphOptions);
+    } else if (allowedInsert) {
+      q.allowGraph(allowedInsert);
+      q.insertGraph(data, this.insertGraphOptions);
+    } else {
+      promise = this._batchInsert(data, params);
     }
-
-    return create(data, params).then(this._commitTransaction(transaction), this._rollbackTransaction(transaction));
+    return promise
+      .then(insertResults => this._getCreatedRecords(insertResults, data, params))
+      .then(this._commitTransaction(transaction), this._rollbackTransaction(transaction))
+      .catch(errorHandler);
   }
 
   /**
@@ -676,7 +725,7 @@ class Service extends AdapterService {
    * @param data
    * @param params
    */
-  _update (id, data, params) {
+  _update (id, data, params = {}) {
     // NOTE : First fetch the item to update to account for user query
     return this._get(id, params)
       .then(() => {
@@ -741,7 +790,7 @@ class Service extends AdapterService {
    * @param data
    * @param params
    */
-  _patch (id, data, params) {
+  _patch (id, data, params = {}) {
     let { filters, query } = this.filterQuery(params);
 
     const allowedUpsert = this.mergeRelations(this.allowedUpsert, params.mergeAllowUpsert);
@@ -814,21 +863,23 @@ class Service extends AdapterService {
         updateKeys(findParams.query);
 
         return q.patch(dataCopy).then(() => {
-          return params.query && params.query.$noSelect ? dataCopy : this._find(findParams).then(page => {
-            const items = page.data || page;
+          return params.query && params.query.$noSelect
+            ? dataCopy
+            : this._find(findParams).then(page => {
+              const items = page.data || page;
 
-            if (id !== null) {
-              if (items.length === 1) {
-                return items[0];
-              } else {
+              if (id !== null) {
+                if (items.length === 1) {
+                  return items[0];
+                } else {
+                  throw new errors.NotFound(`No record found for id '${id}'`);
+                }
+              } else if (!items.length) {
                 throw new errors.NotFound(`No record found for id '${id}'`);
               }
-            } else if (!items.length) {
-              throw new errors.NotFound(`No record found for id '${id}'`);
-            }
 
-            return items;
-          });
+              return items;
+            });
         });
       })
       .catch(errorHandler);
@@ -839,7 +890,7 @@ class Service extends AdapterService {
    * @param id
    * @param params
    */
-  _remove (id, params) {
+  _remove (id, params = {}) {
     params.query = Object.assign({}, params.query);
 
     // NOTE (EK): First fetch the record so that we can return