+++ test/index.test.js
@@ -15,7 +15,18 @@ import PeopleRoomsCustomIdSeparator from './people-rooms-custom-id-separator';
 import Company from './company';
 import Employee from './employee';
 import Client from './client';
-import { Model, UniqueViolationError } from 'objection';
+import {
+  CheckViolationError,
+  ConstraintViolationError,
+  DataError,
+  DBError,
+  ForeignKeyViolationError,
+  Model,
+  NotFoundError,
+  NotNullViolationError,
+  UniqueViolationError,
+  ValidationError
+} from 'objection';
 
 const testSuite = adapterTests([
   '.options',
@@ -30,21 +41,28 @@ const testSuite = adapterTests([
   '.get + $select',
   '.get + id + query',
   '.get + NotFound',
+  '.get + id + query id',
   '.find',
   '.remove',
   '.remove + $select',
   '.remove + id + query',
+  // '.remove + id + query id', // TODO: check why it's failing
   '.remove + multi',
   '.update',
   '.update + $select',
   '.update + id + query',
+  '.update + id + query id',
   '.update + NotFound',
+  '.update + query + NotFound',
   '.patch',
   '.patch + $select',
   '.patch + id + query',
+  '.patch + id + query id',
   '.patch multiple',
-  '.patch multi query',
+  '.patch multi query same',
+  '.patch multi query changed',
   '.patch + NotFound',
+  '.patch + query + NotFound',
   '.create',
   '.create + $select',
   '.create multi',
@@ -325,205 +343,101 @@ describe('Feathers Objection Service', () => {
       }
     });
 
-    describe('SQLite', () => {
-      it('Unknown error code', () => {
+    describe('Error Mappings', () => {
+      it('Unknown error', () => {
         const error = new Error();
-        error.code = 'SQLITE_ERROR';
-        error.errno = 999;
         expect(errorHandler.bind(null, error)).to.throw(errors.GeneralError);
       });
 
-      it('BadRequest 1', () => {
-        const error = new Error();
-        error.code = 'SQLITE_ERROR';
-        error.errno = 1;
-        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
+      it('Validation error', () => {
+        const validationErrTypes = ['ModelValidation', 'RelationExpression', 'UnallowedRelation', 'InvalidGraph', 'unknown-thing'];
+        for (const type of validationErrTypes) {
+          const error = new ValidationError({ type });
+          expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
+        }
       });
 
-      it('BadRequest 8', () => {
-        const error = new Error();
-        error.code = 'SQLITE_ERROR';
-        error.errno = 8;
-        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
-      });
-
-      it('BadRequest 18', () => {
-        const error = new Error();
-        error.code = 'SQLITE_ERROR';
-        error.errno = 18;
-        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
-      });
-
-      it('BadRequest 19', () => {
-        const error = new Error();
-        error.code = 'SQLITE_ERROR';
-        error.errno = 19;
-        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
-      });
-
-      it('BadRequest 20', () => {
-        const error = new Error();
-        error.code = 'SQLITE_ERROR';
-        error.errno = 20;
-        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
-      });
-
-      it('Unavailable 2', () => {
-        const error = new Error();
-        error.code = 'SQLITE_ERROR';
-        error.errno = 2;
-        expect(errorHandler.bind(null, error)).to.throw(errors.Unavailable);
-      });
-
-      it('Forbidden 3', () => {
-        const error = new Error();
-        error.code = 'SQLITE_ERROR';
-        error.errno = 3;
-        expect(errorHandler.bind(null, error)).to.throw(errors.Forbidden);
-      });
-
-      it('Forbidden 23', () => {
-        const error = new Error();
-        error.code = 'SQLITE_ERROR';
-        error.errno = 23;
-        expect(errorHandler.bind(null, error)).to.throw(errors.Forbidden);
-      });
-
-      it('NotFound 12', () => {
-        const error = new Error();
-        error.code = 'SQLITE_ERROR';
-        error.errno = 12;
-        expect(errorHandler.bind(null, error)).to.throw(errors.NotFound);
-      });
-    });
-
-    describe('Objection', () => {
-      it('Unknown error code', () => {
-        const error = new Error();
-        error.statusCode = 999;
-        expect(errorHandler.bind(null, error)).to.throw(errors.GeneralError);
-      });
-
-      it('BadRequest 400', () => {
-        const error = new Error();
-        error.statusCode = 400;
-        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
-      });
-
-      it('NotAuthenticated 401', () => {
-        const error = new Error();
-        error.statusCode = 401;
-        expect(errorHandler.bind(null, error)).to.throw(errors.NotAuthenticated);
-      });
-
-      it('PaymentError 402', () => {
-        const error = new Error();
-        error.statusCode = 402;
-        expect(errorHandler.bind(null, error)).to.throw(errors.PaymentError);
-      });
-
-      it('Forbidden 403', () => {
-        const error = new Error();
-        error.statusCode = 403;
-        expect(errorHandler.bind(null, error)).to.throw(errors.Forbidden);
-      });
-
-      it('NotFound 404', () => {
-        const error = new Error();
-        error.statusCode = 404;
+      it('NotFound error', () => {
+        const error = new NotFoundError();
         expect(errorHandler.bind(null, error)).to.throw(errors.NotFound);
       });
 
-      it('MethodNotAllowed 405', () => {
-        const error = new Error();
-        error.statusCode = 405;
-        expect(errorHandler.bind(null, error)).to.throw(errors.MethodNotAllowed);
-      });
+      it('UniqueViolation error', () => {
+        const error = new UniqueViolationError({
+          nativeError: new Error(),
+          client: 'sqlite',
+          table: 'tableName',
+          columns: ['columnName']
+        });
 
-      it('NotAcceptable 406', () => {
-        const error = new Error();
-        error.statusCode = 406;
-        expect(errorHandler.bind(null, error)).to.throw(errors.NotAcceptable);
-      });
-
-      it('Timeout 408', () => {
-        const error = new Error();
-        error.statusCode = 408;
-        expect(errorHandler.bind(null, error)).to.throw(errors.Timeout);
-      });
-
-      it('Conflict 409', () => {
-        const error = new Error();
-        error.statusCode = 409;
         expect(errorHandler.bind(null, error)).to.throw(errors.Conflict);
       });
 
-      it('Unprocessable 422', () => {
-        const error = new Error();
-        error.statusCode = 422;
-        expect(errorHandler.bind(null, error)).to.throw(errors.Unprocessable);
+      it('UniqueViolation error mysql', () => {
+        const error = new UniqueViolationError({
+          nativeError: { sqlMessage: 'test' },
+          client: 'mysql',
+          table: undefined,
+          columns: undefined,
+          constraint: 'test_constraint'
+        });
+
+        expect(errorHandler.bind(null, error)).to.throw(error.Conflict, 'test');
       });
 
-      it('GeneralError 500', () => {
-        const error = new Error();
-        error.statusCode = 500;
+      it('ConstraintViolation error', () => {
+        const error = new ConstraintViolationError({
+          nativeError: new Error(),
+          client: 'sqlite'
+        });
+
+        expect(errorHandler.bind(null, error)).to.throw(errors.Conflict);
+      });
+
+      it('NotNullViolation error', () => {
+        const error = new NotNullViolationError({
+          nativeError: new Error(),
+          client: 'sqlite',
+          column: 'columnName'
+        });
+
+        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
+      });
+
+      it('ForeignKeyViolation error', () => {
+        const error = new ForeignKeyViolationError({
+          nativeError: new Error(),
+          client: 'sqlite'
+        });
+
+        expect(errorHandler.bind(null, error)).to.throw(errors.Conflict);
+      });
+
+      it('CheckViolation error', () => {
+        const error = new CheckViolationError({
+          nativeError: new Error(),
+          client: 'sqlite'
+        });
+
+        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
+      });
+
+      it('Data error', () => {
+        const error = new DataError({
+          nativeError: new Error(),
+          client: 'sqlite'
+        });
+
+        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
+      });
+
+      it('Database error', () => {
+        const error = new DBError({
+          nativeError: new Error(),
+          client: 'sqlite'
+        });
+
         expect(errorHandler.bind(null, error)).to.throw(errors.GeneralError);
-      });
-
-      it('NotImplemented 501', () => {
-        const error = new Error();
-        error.statusCode = 501;
-        expect(errorHandler.bind(null, error)).to.throw(errors.NotImplemented);
-      });
-
-      it('Unavailable 503', () => {
-        const error = new Error();
-        error.statusCode = 503;
-        expect(errorHandler.bind(null, error)).to.throw(errors.Unavailable);
-      });
-    });
-
-    describe('Postgres', () => {
-      it('Unknown error code', () => {
-        const error = new Error();
-        error.code = '999';
-        expect(errorHandler.bind(null, error)).to.throw(errors.GeneralError);
-      });
-
-      it('Forbidden 28', () => {
-        const error = new Error();
-        error.code = '28';
-        expect(errorHandler.bind(null, error)).to.throw(errors.Forbidden);
-      });
-
-      it('Forbidden 42', () => {
-        const error = new Error();
-        error.code = '42';
-        expect(errorHandler.bind(null, error)).to.throw(errors.Forbidden);
-      });
-
-      it('BadRequest 20', () => {
-        const error = new Error();
-        error.code = '20';
-        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
-      });
-
-      it('BadRequest 21', () => {
-        const error = new Error();
-        error.code = '21';
-        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
-      });
-
-      it('BadRequest 22', () => {
-        const error = new Error();
-        error.code = '22';
-        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
-      });
-
-      it('BadRequest 23', () => {
-        const error = new Error();
-        error.code = '23';
-        expect(errorHandler.bind(null, error)).to.throw(errors.BadRequest);
       });
     });
   });
@@ -729,6 +643,20 @@ describe('Feathers Objection Service', () => {
         return peopleRooms.find().then(data => {
           expect(data.length).to.equal(2);
         });
+      });
+    });
+
+    it('allows get queries with $select', () => {
+      return peopleRooms.get([2, 2], { query: { $select: ['admin'] } }).then(data => {
+        expect(data.admin).to.equal(1);
+      });
+    });
+
+    it('allows find queries with $select', () => {
+      return peopleRooms.find({ query: { roomId: 2, $select: ['admin'] } }).then(data => {
+        expect(data.length).to.equal(2);
+        expect(data[0].admin).to.equal(0);
+        expect(data[1].admin).to.equal(1);
       });
     });
   });
@@ -1135,9 +1063,11 @@ describe('Feathers Objection Service', () => {
     });
 
     it('allows upsertGraph queries on update', () => {
-      const newClients = (google.clients) ? google.clients.concat([{
-        name: 'Ken Patrick'
-      }]) : [];
+      const newClients = (google.clients)
+        ? google.clients.concat([{
+            name: 'Ken Patrick'
+          }])
+        : [];
 
       return companies
         .update(google.id, {
@@ -1153,9 +1083,11 @@ describe('Feathers Objection Service', () => {
     });
 
     it('forbid upsertGraph if data do not match update item', () => {
-      const newClients = (google.clients) ? google.clients.concat([{
-        name: 'Ken Patrick'
-      }]) : [];
+      const newClients = (google.clients)
+        ? google.clients.concat([{
+            name: 'Ken Patrick'
+          }])
+        : [];
 
       return companies
         .update(apple.id, {
@@ -1171,9 +1103,11 @@ describe('Feathers Objection Service', () => {
     });
 
     it('forbid upsertGraph if data do not match patch item', () => {
-      const newClients = (google.clients) ? google.clients.concat([{
-        name: 'Ken Patrick'
-      }]) : [];
+      const newClients = (google.clients)
+        ? google.clients.concat([{
+            name: 'Ken Patrick'
+          }])
+        : [];
 
       return companies
         .patch(apple.id, {
@@ -1189,9 +1123,11 @@ describe('Feathers Objection Service', () => {
     });
 
     it('allows upsertGraph queries on patch', () => {
-      const newClients = (google.clients) ? google.clients.concat([{
-        name: 'Ken Patrick'
-      }]) : [];
+      const newClients = (google.clients)
+        ? google.clients.concat([{
+            name: 'Ken Patrick'
+          }])
+        : [];
 
       return companies
         .patch(google.id, {
@@ -1210,9 +1146,11 @@ describe('Feathers Objection Service', () => {
     });
 
     it('allows upsertGraph queries on patch with query param', () => {
-      const newClients = (google.clients) ? google.clients.concat([{
-        name: 'Ken Patrick'
-      }]) : [];
+      const newClients = (google.clients)
+        ? google.clients.concat([{
+            name: 'Ken Patrick'
+          }])
+        : [];
 
       return companies
         .patch(google.id, {
@@ -1555,7 +1493,7 @@ describe('Feathers Objection Service', () => {
       }).catch(function (error) {
         expect(error).to.be.ok;
         expect(error instanceof errors.GeneralError).to.be.ok;
-        expect(error.message).to.equal('select `companies`.* from `companies` where CAST(`companies`.`jsonObject`#>>\'{numberField}\' AS text) = 1.5 - SQLITE_ERROR: unrecognized token: "#"');
+        expect(error[ERROR].message).to.equal('select `companies`.* from `companies` where CAST(`companies`.`jsonObject`#>>\'{numberField}\' AS text) = 1.5 - SQLITE_ERROR: unrecognized token: "#"');
       });
     });
 
@@ -1565,7 +1503,7 @@ describe('Feathers Objection Service', () => {
       }).catch(function (error) {
         expect(error).to.be.ok;
         expect(error instanceof errors.GeneralError).to.be.ok;
-        expect(error.message).to.equal('select `companies`.* from `companies` where CAST(`companies`.`jsonObject`#>>\'{numberField}\' AS text) > 1.5 - SQLITE_ERROR: unrecognized token: "#"');
+        expect(error[ERROR].message).to.equal('select `companies`.* from `companies` where CAST(`companies`.`jsonObject`#>>\'{numberField}\' AS text) > 1.5 - SQLITE_ERROR: unrecognized token: "#"');
       });
     });
 
@@ -1575,7 +1513,7 @@ describe('Feathers Objection Service', () => {
       }).catch(function (error) {
         expect(error).to.be.ok;
         expect(error instanceof errors.GeneralError).to.be.ok;
-        expect(error.message).to.equal('select `companies`.* from `companies` where CAST(`companies`.`jsonObject`#>>\'{objectField,object}\' AS text) = \'string in jsonObject.objectField.object\' - SQLITE_ERROR: unrecognized token: "#"');
+        expect(error[ERROR].message).to.equal('select `companies`.* from `companies` where CAST(`companies`.`jsonObject`#>>\'{objectField,object}\' AS text) = \'string in jsonObject.objectField.object\' - SQLITE_ERROR: unrecognized token: "#"');
       });
     });
 
@@ -1591,7 +1529,7 @@ describe('Feathers Objection Service', () => {
       }).catch(function (error) {
         expect(error).to.be.ok;
         expect(error instanceof errors.GeneralError).to.be.ok;
-        expect(error.message).to.equal('select `companies`.* from `companies` where CAST(`companies`.`jsonArray`#>>\'{0,objectField,object}\' AS text) = \'I\'\'m string in jsonArray[0].objectField.object\' - SQLITE_ERROR: unrecognized token: "#"');
+        expect(error[ERROR].message).to.equal('select `companies`.* from `companies` where CAST(`companies`.`jsonArray`#>>\'{0,objectField,object}\' AS text) = \'I\'\'m string in jsonArray[0].objectField.object\' - SQLITE_ERROR: unrecognized token: "#"');
       });
     });
   });
@@ -1895,8 +1833,8 @@ describe('Feathers Objection Service', () => {
     it('Rollback on sub insert failure', () => {
       // Dan Davis already exists
       return companies.create({ name: 'Compaq', clients: [{ name: 'Dan Davis' }] }, { atomic: true }).catch((error) => {
-        expect(error instanceof errors.GeneralError).to.be.ok;
-        expect(error.message).to.match(/SQLITE_CONSTRAINT: UNIQUE/);
+        expect(error instanceof errors.Conflict).to.be.ok;
+        expect(error[ERROR].message).to.match(/SQLITE_CONSTRAINT: UNIQUE/);
         return companies.find({ query: { name: 'Compaq', $eager: 'clients' } }).then(
           (data) => {
             expect(data.length).to.equal(0);
@@ -1907,8 +1845,8 @@ describe('Feathers Objection Service', () => {
     it('Rollback on multi insert failure', () => {
       // Google already exists
       return companies.create([{ name: 'Google' }, { name: 'Compaq' }], { atomic: true }).catch((error) => {
-        expect(error instanceof errors.GeneralError).to.be.ok;
-        expect(error.message).to.match(/SQLITE_CONSTRAINT: UNIQUE/);
+        expect(error instanceof errors.Conflict).to.be.ok;
+        expect(error[ERROR].message).to.match(/SQLITE_CONSTRAINT: UNIQUE/);
         return companies.find({ query: { name: 'Compaq' } }).then(
           (data) => {
             expect(data.length).to.equal(0);
@@ -1933,8 +1871,8 @@ describe('Feathers Objection Service', () => {
             }
           ]
         }, { atomic: true }).catch((error) => {
-          expect(error instanceof errors.GeneralError).to.be.ok;
-          expect(error.message).to.match(/SQLITE_CONSTRAINT: UNIQUE/);
+          expect(error instanceof errors.Conflict).to.be.ok;
+          expect(error[ERROR].message).to.match(/SQLITE_CONSTRAINT: UNIQUE/);
           return companies.find({ query: { name: 'Google', $eager: 'clients' } }).then(
             (data) => {
               expect(data.length).to.equal(1);
@@ -2197,6 +2135,49 @@ describe('Feathers Objection Service', () => {
         expect(data.data[0].name).to.be.equal('Google');
         expect(data.data[0].employees.length).to.be.equal(1);
         expect(data.data[0].employees[0].name).to.be.equal('John');
+      });
+    });
+
+    it('params.modifierFiltersResults=false does not apply count from modify query', () => {
+      companies.options.paginate = {
+        default: 2,
+        max: 2
+      };
+
+      return companies.find({
+        query: { $modify: ['withRelationAndGroupBy'] }, modifierFiltersResults: false
+      }).then(data => {
+        expect(data.total).to.be.equal(2);
+        expect(data.data.length).to.be.equal(2);
+        expect(data.data[0].name).to.be.equal('Google');
+      });
+    });
+
+    it('params.modifierFiltersResults=true applies count from modify query', () => {
+      companies.options.paginate = {
+        default: 2,
+        max: 2
+      };
+
+      return companies.find({
+        query: { $modify: ['withRelationAndGroupBy'] }, modifierFiltersResults: true
+      }).then(data => {
+        expect(data.total).to.be.equal(1); // count result from GROUP BY
+        expect(data.data.length).to.be.equal(2);
+        expect(data.data[0].name).to.be.equal('Google');
+      });
+    });
+
+    it('params.modifierFiltersResults=undefined applies count from modify query', () => {
+      companies.options.paginate = {
+        default: 2,
+        max: 2
+      };
+
+      return companies.find({ query: { $modify: ['withRelationAndGroupBy'] } }).then(data => {
+        expect(data.total).to.be.equal(1); // count result from GROUP BY
+        expect(data.data.length).to.be.equal(2);
+        expect(data.data[0].name).to.be.equal('Google');
       });
     });